Split v2.py into multiple modules
