+++ vastdb/__init__.py
@@ -1,0 +1,4 @@
+from . import session
+
+def connect(*args, **kw):
+    return session.Session(*args, **kw)
