+++ vastdb/session.py
@@ -1,0 +1,28 @@
+from . import internal_commands
+from . import transaction
+
+import boto3
+
+import os
+
+
+class Session:
+    def __init__(self, access=None, secret=None, endpoint=None):
+        if access is None:
+            access = os.environ['AWS_ACCESS_KEY_ID']
+        if secret is None:
+            secret = os.environ['AWS_SECRET_ACCESS_KEY']
+        if endpoint is None:
+            endpoint = os.environ['AWS_S3_ENDPOINT_URL']
+
+        self.api = internal_commands.VastdbApi(endpoint, access, secret)
+        self.s3 = boto3.client('s3',
+            aws_access_key_id=access,
+            aws_secret_access_key=secret,
+            endpoint_url=endpoint)
+
+    def __repr__(self):
+        return f'{self.__class__.__name__}(endpoint={self.api.url}, access={self.api.access_key})'
+
+    def transaction(self):
+        return transaction.Transaction(self)
