+++ vastdb/util.py
@@ -4,8 +4,9 @@ from typing import Callable
 import pyarrow as pa
 import pyarrow.parquet as pq
 
-from vastdb.v2 import Table, Schema
-from vastdb.errors import InvalidArgumentError
+from .errors import InvalidArgumentError
+from .schema import Schema
+from .table import Table
 
 log = logging.getLogger(__name__)
